routers are working, just need to specify attachment of id to show page
