+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^1.3.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
